requirements.txt
appnope==0.1.4
arch==8.0.0
asttokens==3.0.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
comm==0.2.3
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
fonttools==4.60.1
frozendict==2.4.6
hurst==0.0.5
idna==3.10
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.7
matplotlib-inline==0.1.7
multitasking==0.0.12
nest-asyncio==1.6.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pandas-datareader==0.10.0
parso==0.8.5
patsy==1.0.2
peewee==3.18.2
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
statsmodels==0.14.5
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
websockets==15.0.1
yfinance==0.2.66
